api: restructured Flask code to modularise it
