+++ api/test/pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 addopts = --strict-markers
 markers =
-    slow: marks tests as slow (deselect with '-m "not slow"')
+	slow: marks tests as slow (deselect with '-m "not slow"')
